feat(node): persist the chain and accept signed transactions
The node rebuilt an empty chain at every start and offered no way to move
value: add_block was unreachable from outside the process.

- ChainStore loads the chain from CHAIN_PATH at boot and writes it through
  after every accepted block, so a restart resumes from the same tip.
- NodeIdentity loads (or generates, 0600) the ed25519 seal key from
  AUTHORITY_KEY_PATH and registers it as the chain authority.
- POST /transactions takes a client-signed transfer, seals it into a block
  and persists it; POST /accounts registers the key a sender is verified
  against; GET /accounts/<name> returns the balance and the next nonce a
  client needs to sign. Failures answer with a typed JSON error and the
  right status instead of a 500.
- The chain loaded from disk is validated before the node serves anything.
- Rocket assembly moved to server::build so integration tests can drive the
  real routes; gRPC amounts follow the ledger to uint64.
- .env.example documents every variable, including the new ones.

Minting stays internal: transactions from the System account are refused
at the API boundary.

+++ zion-core/src/api/grpc.rs
@@ -30,7 +30,7 @@ impl TransactionService for MyTransactionService {
                 id,
                 sender: tx.sender,
                 recipient: tx.receiver,
-                amount: tx.amount as f32,
+                amount: tx.amount,
             })),
             None => Err(Status::not_found("Transaction not found")),
         }
